requirements.txt
starlette
pytest
httpx
pydantic
fastapi
uvicorn
psycopg2-binary
devtools